Reverting points to int32 due to open cv polylines issue

+++ gyakujinton/Shape/Shape.py
@@ -2,4 +2,4 @@ class Shape():
     def __new__(cls, points):
         import numpy as np
 
-        return np.array([points], np.float32)
+        return np.array([points], np.int32)
